Rename gitlab/pj to gitlab/project
